Fixed command & base template.

+++ src/AppBundle/Command/RemoveZipFileCommand.php
@@ -5,6 +5,7 @@ namespace AppBundle\Command;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Security\Csrf\TokenStorage\TokenStorageInterface;
 
 class RemoveZipFileCommand extends ContainerAwareCommand
 {
@@ -17,11 +18,13 @@ class RemoveZipFileCommand extends ContainerAwareCommand
 	protected function execute(InputInterface $input, OutputInterface $output)
 	{
 		$container = $this->getContainer();
-		$user = $container->get('security.token_storage')
-			->getToken()
-			->getUser();
+		$token = $container->get('security.token_storage')->getToken();
 
-		$file = sprintf("%s/%s.zip", $container->getParameter('library'), $user->getId());
+		if(false === $token instanceof TokenStorageInterface) {
+			return;
+		}
+
+		$file = sprintf("%s/%s.zip", $container->getParameter('library'), $token->getUser()->getId());
 
 		if(false !== file_exists($file)) {
 			unlink($file);
